+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Unreleased
+
+- Removed `JsException` unpickle special case. This was fixed by
+  [pyodide/pyodide#3387](https://github.com/pyodide/pyodide/pull/3387).
+  [#91](https://github.com/pyodide/pytest-pyodide/pull/91)
+
+- Dropped support for Pyodide version `0.21.x`.
+  [#91](https://github.com/pyodide/pytest-pyodide/pull/91)
+
 ## [0.51.0] - 2023-05-10
 
 - Added test templates files in the package.